# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: E-commerce shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, preparing, delivered, cancelled]
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for customer or admin
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send Telegram notification to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationBody"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /telegram-webhook:
    post:
      operationId: telegramWebhook
      tags: [notifications]
      summary: Telegram bot webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        oldPrice:
          type: number
          nullable: true
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        inStock:
          type: boolean
        isLiked:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - price
        - images
        - inStock
        - isLiked
        - createdAt
    ProductsListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - products
        - total
        - page
        - limit
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        oldPrice:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        inStock:
          type: boolean
      required:
        - name
        - price
        - images
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        oldPrice:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        inStock:
          type: boolean
...
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, preparing, delivered, cancelled]
        deliveryMethod:
          type: string
          enum: [delivery, pickup]
        paymentMethod:
          type: string
          enum: [cash, card, online]
        address:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        totalPrice:
          type: number
        deliveryFee:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required:
        - id
        - customerId
        - status
        - deliveryMethod
        - paymentMethod
        - totalPrice
        - deliveryFee
        - items
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
          nullable: true
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - price
    CreateOrderBody:
      type: object
      properties:
        deliveryMethod:
          type: string
          enum: [delivery, pickup]
        paymentMethod:
          type: string
          enum: [cash, card, online]
        address:
          type: string
        note:
          type: string
      required:
        - deliveryMethod
        - paymentMethod
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, preparing, delivered, cancelled]
      required:
        - status
...
    Message:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        senderType:
          type: string
          enum: [customer, admin]
        text:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - customerId
        - senderType
        - text
        - isRead
        - createdAt
    SendMessageBody:
      type: object
      properties:
        customerId:
          type: integer
        text:
          type: string
        senderType:
          type: string
          enum: [customer, admin]
      required:
        - text
        - senderType
...
    SendNotificationBody:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...