# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce shop API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyDeliveryMethod } from "./createOrderBodyDeliveryMethod";
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";

export interface CreateOrderBody {
  deliveryMethod: CreateOrderBodyDeliveryMethod;
  paymentMethod: CreateOrderBodyPaymentMethod;
  address?: string;
  note?: string;
}