# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDeliveryMethod } from "./orderDeliveryMethod";
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerId: number;
  customerName?: string | null;
  customerPhone?: string | null;
  status: OrderStatus;
  deliveryMethod: OrderDeliveryMethod;
  paymentMethod: OrderPaymentMethod;
  address?: string | null;
  note?: string | null;
  totalPrice: number;
  deliveryFee: number;
  items: OrderItem[];
  createdAt: string;
}